Refactor scrapers to clarify interface, remove CITYOFBOSTON-specific logic from Base.
